+++ app/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -2,8 +2,8 @@ package nextstep.jwp.controller;
 
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
-import nextstep.jwp.http.ContentType;
-import nextstep.jwp.http.FileReader;
+import nextstep.jwp.utils.ContentType;
+import nextstep.jwp.utils.FileReader;
 import nextstep.jwp.http.HttpRequest;
 import nextstep.jwp.http.HttpResponse;
 import nextstep.jwp.http.HttpStatus;
