+++ app/src/main/java/nextstep/jwp/http/HttpResponse.java
@@ -2,6 +2,7 @@ package nextstep.jwp.http;
 
 import java.io.IOException;
 import java.io.OutputStream;
+import nextstep.jwp.utils.ContentType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
