+++ app/src/main/java/nextstep/jwp/utils/ContentType.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.utils;
 
 import java.util.Arrays;
