+++ app/src/main/java/nextstep/jwp/utils/FileReader.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.http;
+package nextstep.jwp.utils;
 
 import java.io.File;
 import java.io.IOException;
